requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-braces==1.14.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-oauth-toolkit==1.3.2
django-parcel-ssr==0.7.0
django-rest-auth==0.9.5
django-heroku==0.3.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==4.2.0
gcloud==0.18.3
googleapis-common-protos==1.51.0
h2==3.2.0
hiredis==1.0.1
hpack==3.0.0
httplib2==0.17.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jws==0.1.3
Markdown==3.2.1
MouseInfo==0.1.2
msgpack==0.6.2
msgpack-python==0.5.6
numpy==1.18.1
oauth2client==4.1.2
oauthlib==3.1.0
opencv-python==4.2.0.32
Pillow==7.0.0
priority==1.3.0
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
pycparser==2.20
pycryptodome==3.9.7
PyGetWindow==0.0.8
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMsgBox==1.0.7
pyOpenSSL==19.1.0
pyperclip==1.7.0
#pypiwin32==223
Pyrebase4==4.3.0
PyRect==0.1.4
PyScreeze==0.1.26
python-jwt==2.0.1
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2019.3
#pywin32==227
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rest-social-auth==4.1.0
rsa==4.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.3.1
urllib3==1.25.8
websocket-client==0.57.0
whitenoise==5.0.1
zope.interface==5.0.1
gunicorn==19.7.1
whitenoise==5.0.1
python-decouple==3.3
dj-database-url==0.5.0
python-dotenv==0.14.0